+++ backend/backend/src/main/java/de/morninggrind/backend/domain/OrderItem.java
@@ -1,13 +1,15 @@
 package de.morninggrind.backend.domain;
 
 import lombok.Data;
+
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Data
 public class OrderItem {
     private UUID id;
     private int quantity;
-    private float priceAtPurchase;
+    private BigDecimal priceAtPurchase;
     private Long orderId;
     private Long productId;
 }
